requirements.txt
langchain_text_splitters
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.40
langchain-google-genai==1.0.10
python-dotenv
pypdf
langchain-postgres